Sharing a whole bunch of code for options between ReSharper versions.
