+++ Project/Src/AddIns/ReSharper710/CodeCleanup/Options/MaintainabilityOptions.cs
@@ -19,10 +19,6 @@ namespace StyleCop.ReSharper710.CodeCleanup.Options
 {
     #region Using Directives
 
-    using System.ComponentModel;
-    using System.Reflection;
-    using System.Text;
-
     using ReSharperBase.CodeCleanup.Options;
 
     using StyleCop.ReSharper710.Core;
@@ -32,7 +28,7 @@ namespace StyleCop.ReSharper710.CodeCleanup.Options
     /// <summary>
     /// Defines options for MaintainabilityOptions.
     /// </summary>
-    public class MaintainabilityOptions : OptionsBase
+    public class MaintainabilityOptions : MaintainabilityOptionsBase
     {
         #region Constructors and Destructors
 
@@ -41,107 +37,7 @@ namespace StyleCop.ReSharper710.CodeCleanup.Options
         /// </summary>
         public MaintainabilityOptions()
         {
-            base.InitPropertiesDefaults(Utils.GetStyleCopSettings());
-        }
-
-        #endregion
-
-        #region Public Properties
-
-        /// <summary>
-        /// Gets or sets a value indicating whether SA1119StatementMustNotUseUnnecessaryParenthesis.
-        /// </summary>
-        [DisplayName("1119: Statement Must Not Use Unnecessary Parenthesis")]
-        public bool SA1119StatementMustNotUseUnnecessaryParenthesis { get; set; }
-
-        #endregion
-
-        #region Properties
-
-        /// <summary>
-        /// Gets the name of the analyzer.
-        /// </summary>
-        protected override string AnalyzerName
-        {
-            get
-            {
-                return "StyleCop.CSharp.MaintainabilityRules";
-            }
-        }
-
-        #endregion
-
-        #region Public Methods and Operators
-
-        /// <summary>
-        /// Returns a concatenated summary of the current options settings.
-        /// </summary>
-        /// <returns>
-        /// A String of the options.
-        /// </returns>
-        public override string ToString()
-        {
-            StringBuilder sb = new StringBuilder();
-            PropertyInfo[] properties = this.GetType().GetProperties();
-
-            for (int i = 0; i < properties.Length; i++)
-            {
-                PropertyInfo property = properties[i];
-                if (i > 0)
-                {
-                    sb.Append(", ");
-                }
-
-                sb.Append(this.GetPropertyDecription(property));
-            }
-
-            return sb.ToString();
-        }
-
-        #endregion
-
-        #region Methods
-
-        /// <summary>
-        /// Builds a string representation of the property value.
-        /// </summary>
-        /// <param name="propertyInfo">
-        /// The propertyInfo to build the description for.
-        /// </param>
-        /// <returns>
-        /// The string representation.
-        /// </returns>
-        private string GetPropertyDecription(PropertyInfo propertyInfo)
-        {
-            string propertyValue = propertyInfo.GetValue(this, null).ToString();
-
-            string propName = string.Empty;
-            string propValue = string.Empty;
-            DisplayNameAttribute[] displayNameAttributes = (DisplayNameAttribute[])propertyInfo.GetCustomAttributes(typeof(DisplayNameAttribute), false);
-            if (displayNameAttributes.Length == 1)
-            {
-                propName = displayNameAttributes[0].DisplayName;
-            }
-
-            if (propertyInfo.PropertyType == typeof(bool))
-            {
-                propValue = propertyValue == "True" ? "Yes" : "No";
-            }
-            else
-            {
-                FieldInfo field = propertyInfo.PropertyType.GetField(propertyValue);
-
-                if (field != null)
-                {
-                    DescriptionAttribute[] descriptionAttributes = (DescriptionAttribute[])field.GetCustomAttributes(typeof(DescriptionAttribute), false);
-                    if (descriptionAttributes.Length == 1)
-                    {
-                        propValue = descriptionAttributes[0].Description;
-                    }
-                }
-            }
-
-            return string.Format("{0} = {1}", propName, propValue);
+            this.InitPropertiesDefaults(Utils.GetStyleCopSettings());
         }
 
         #endregion
